Made Contact Page with suggestion form
Contact page now includes a remote suggestion form
Added ajax handlers to deal with errors

Wrote server-side validation for the form, data sent back as json.
Generated a mailer which for now sends me an email with the suggestion.

ajax:success handler deals with errors and notifies user.
Fields flash if there is an error.

Form is verified with recaptcha.

Right now it only does server-side validation, and the errors get
sent back to the client as json and handled.
To-Do: Also do client-side validation, write ajax:beforeSend handler
This will cut down with superfluous requests from the client