+Many new features added to Resumable
A new set of client-side scripts keeps track of changes made
to the form and monitors browser close events.
If a change is made and the tab is closed, the user is prompted and
the current state of the form is saved via ajax on the server-side.

If the user has any saved forms, a new tab shows up on the application
page. At the moment this tab is titled "Resumable" but maybe I should
call it "Recovery" or something else more urgent sounding.

This resumable tab shows the number of things that require the user's attention
and leads to a page with a list of resumable states.
Clicking on a state takes the user to the corresponding form and JS hackery
fills in the form by parsing a special formdata parameter to the request.

This is probably not the best way to do it, but it is general enough so that it
will work on any form throughout the application. Rails' form_for doesn't leave
any room for partial saves, so this avoids the server-side all together.