Minor but important bug fixes
The resumable save-state links work perfectly fine
from a normal path like /corpora/1/edit
but what if a save-state occurs on a form that just came
from a save-state?
Well in that case, tacking on a formdata parameter to the request
doesn't work very well, and we have an exponentially increasing
url (it keeps doubling).

The fix: store the url in a less naive way, if there are parameters parse them out.

To-Do: current fix uses regex, switch to indexOf('?') and subtr because
it's faster.