+Integrated Resumable into Publication CRUD
Resumable uploader server-side is set up
to use a single folder, instead of the old uploader's
nested folder structure. So, files are put together with
the "unique" resumable-id prepended to the filename.

This is a solution for multiple people uploading things which
may or may not conflict in name. Previously it did not have
to be done because each upload was in it's own subfolder
with a UUID.

Problem: If we naively use the resumable uploader
in the same manner as the previous uploader is used,
as in we move the target file from the upload directory
then we have large obscure filenames (with the resumable-id tacked on)

There are two solutions:
 -Modify the resumable controller to also include in sessions the original
  filename for any other controller to use if necessary.
 -Ignore the original filename. Move the file from the upload
  directory as an modified form of the resource name
  (i.e if publication, use publication title,
   if corpus then corpus name, etc)

Current solution uses the second method.

Minor issue with this: What happens when you rename the
resource? Let's say I rename the Publication.
Should I also rename the associated pdf?
(Not a huge problem, ignorable for now, resource names
are not expected to change very often...)